% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonCreate}
\alias{HIRMonCreate}
\title{Make an HIRMon object from a transactions data frame}
\usage{
HIRMonCreate(
  hirObj,
  data = HIRMonTakeData(hirObj),
  hubColumnName = names(data)[[1]],
  itemColumnName = names(data)[[2]],
  valueColumnName = NULL,
  addTagTypesToColumnNamesQ = FALSE,
  sep = ":"
)
}
\arguments{
\item{hirObj}{An HIRMon object.}

\item{data}{The hub-item transactions data frame.}

\item{hubColumnName}{The name of the column with the hub tags.}

\item{itemColumnName}{The name of the column with the item tags.}

\item{valueColumnName}{The name of the column with the hub-item associating values.
If NULL only the columns \code{hubColumnName} and \code{itemColumnName} are used.}

\item{addTagTypesToColumnNamesQ}{Should the tag types be added as prefixes
to the column names of the corresponding sub-matrices?}

\item{sep}{Separator for the prefixes of the columns names.}
}
\value{
HIR object.
}
\description{
Creates a sparse matrix recommender from transactions data and a list of tag types,
and makes is an HIRMon object.
}
\details{
An S3 object is returned that is list with class attribute set to "HIR".
Calls \code{\link{HIRMonCreateFromMatrices}}.
}
\seealso{
Other Creation functions: 
\code{\link{HIRMonCreateFromMatrices}()}
}
\concept{Creation functions}
