% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonCreateFromMatrices}
\alias{HIRMonCreateFromMatrices}
\title{Creation an HIRMon object with a list of matrices.}
\usage{
HIRMonCreateFromMatrices(
  hirObj,
  hubItemMatrix,
  itemHubMatrix = t(hubItemMatrix),
  columnStochasticQ = TRUE,
  hubColumnName = "Hub",
  itemColumnName = "Item",
  addTagTypesToColumnNamesQ = FALSE,
  sep = ":"
)
}
\arguments{
\item{hirObj}{An HIRMon object.}

\item{hubItemMatrix}{A hub-item matrix.}

\item{itemHubMatrix}{An item-hub matrix.
If NULL same as \code{t(hubItemMatrix)}.}

\item{columnStochasticQ}{Should the bi-partite matrix be column stochastic or not?}

\item{hubColumnName}{The column name of recommender hubs (in data and recommendations).}

\item{itemColumnName}{The column name of recommender items (in data and recommendations).}

\item{addTagTypesToColumnNamesQ}{Should the tag types be added as prefixes
to the column names of the corresponding sub-matrices?}

\item{sep}{Separator for the prefixes of the columns names.}
}
\value{
An HIRMon object.
}
\description{
Creates a sparse matrix recommender from a list of matrices and a corresponding list of tag types,
and makes is an HIRMon object.
}
\details{
An S3 object is returned that is list with class attribute set to "SMR".
}
\seealso{
Other Creation functions: 
\code{\link{HIRMonCreate}()}
}
\concept{Creation functions}
