% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonEcho}
\alias{HIRMonEcho}
\title{Echo.}
\usage{
HIRMonEcho(hirObj, x)
}
\arguments{
\item{hirObj}{An HIRMon object.}

\item{x}{Object to be echoed.}
}
\value{
An HIRMon object
}
\description{
Echoes given argument.
}
\concept{Echo functions}
