% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonEchoFunctionValue}
\alias{HIRMonEchoFunctionValue}
\title{Echo function application to monad's value.}
\usage{
HIRMonEchoFunctionValue(hirObj, f)
}
\arguments{
\item{hirObj}{An HIRMon object.}

\item{f}{A function to be applied to \code{hirObj$Value}.}
}
\value{
A HIRMon object.
}
\description{
Applies a function to the "Value" element/member of the monad object
and prints the result.
}
\details{
Prints \code{f(hirObj$Value)}.
}
