% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonEchoValue}
\alias{HIRMonEchoValue}
\title{Echo monad's value.}
\usage{
HIRMonEchoValue(hirObj)
}
\arguments{
\item{hirObj}{An HIRMon object.}
}
\value{
A HIRMon object.
}
\description{
Prints the "Value" element/member of the monad object.
}
\details{
Prints \code{hirObj$Value}.
}
