% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonFailure}
\alias{HIRMonFailure}
\title{Failure producing function}
\usage{
HIRMonFailure(hirObj, message = NULL)
}
\arguments{
\item{hirObj}{An HIRMon object.}

\item{message}{A message to echoed. If NULL no message is echoed.}
}
\value{
A HIRMon object.
}
\description{
Puts the monad object into a monadic failure.
}
