% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\docType{data}
\name{HIRMonFailureSymbol}
\alias{HIRMonFailureSymbol}
\title{Failure symbol for HIRMon.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
HIRMonFailureSymbol
}
\description{
Failure symbol for the monad HIRMon.
}
\keyword{datasets}
