% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonOption}
\alias{HIRMonOption}
\title{Optional function application to monad's object.}
\usage{
HIRMonOption(hirObj, f)
}
\arguments{
\item{hirObj}{An HIRMon object.}

\item{f}{A function to be applied to the monad object.}
}
\value{
A HIRMon object.
}
\description{
If monadic failure is obtained from \code{f(hirObj)}
then returns the original \code{hirObj};
else returns the result of \code{f(hirObj)}.
}
\details{
In general \code{f} should return a monad object,
but that is not enforced.
}
