% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonSetM}
\alias{HIRMonSetM}
\title{Set M.}
\usage{
HIRMonSetM(hirObj, M)
}
\arguments{
\item{hirObj}{An HIRMon object.}

\item{M}{An object member to be set.}
}
\value{
An HIRMon object.
}
\description{
Sets M into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{HIRMonSetBiasVector}()},
\code{\link{HIRMonSetData}()},
\code{\link{HIRMonSetHubItemMatrix}()},
\code{\link{HIRMonSetItemHubMatrix}()},
\code{\link{HIRMonSetM01}()},
\code{\link{HIRMonSetTagTypeRanges}()},
\code{\link{HIRMonSetTagTypes}()},
\code{\link{HIRMonSetValue}()},
\code{\link{HIRMonTakeBiasVector}()},
\code{\link{HIRMonTakeData}()},
\code{\link{HIRMonTakeHubItemMatrix}()},
\code{\link{HIRMonTakeItemHubMatrix}()},
\code{\link{HIRMonTakeM01}()},
\code{\link{HIRMonTakeM}()},
\code{\link{HIRMonTakeTagTypeRanges}()},
\code{\link{HIRMonTakeTagTypes}()},
\code{\link{HIRMonTakeValue}()}
}
\concept{Set/Take functions}
