% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonSetM01}
\alias{HIRMonSetM01}
\title{Set M01.}
\usage{
HIRMonSetM01(hirObj, M01)
}
\arguments{
\item{hirObj}{An HIRMon object.}

\item{M01}{An object member to be set.}
}
\value{
An HIRMon object.
}
\description{
Sets M01 into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{HIRMonSetBiasVector}()},
\code{\link{HIRMonSetData}()},
\code{\link{HIRMonSetHubItemMatrix}()},
\code{\link{HIRMonSetItemHubMatrix}()},
\code{\link{HIRMonSetM}()},
\code{\link{HIRMonSetTagTypeRanges}()},
\code{\link{HIRMonSetTagTypes}()},
\code{\link{HIRMonSetValue}()},
\code{\link{HIRMonTakeBiasVector}()},
\code{\link{HIRMonTakeData}()},
\code{\link{HIRMonTakeHubItemMatrix}()},
\code{\link{HIRMonTakeItemHubMatrix}()},
\code{\link{HIRMonTakeM01}()},
\code{\link{HIRMonTakeM}()},
\code{\link{HIRMonTakeTagTypeRanges}()},
\code{\link{HIRMonTakeTagTypes}()},
\code{\link{HIRMonTakeValue}()}
}
\concept{Set/Take functions}
