% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonSetValue}
\alias{HIRMonSetValue}
\title{Set the value in a HIRMon object.}
\usage{
HIRMonSetValue(hirObj, value)
}
\arguments{
\item{hirObj}{An HIRMon object.}

\item{value}{The new value.}
}
\value{
A HIRMon object.
}
\description{
Sets the value in a HIRMon monad object.
}
\details{
Assigns \code{value} to \code{hirObj$Value}.
}
\seealso{
Other Set/Take functions: 
\code{\link{HIRMonSetBiasVector}()},
\code{\link{HIRMonSetData}()},
\code{\link{HIRMonSetHubItemMatrix}()},
\code{\link{HIRMonSetItemHubMatrix}()},
\code{\link{HIRMonSetM01}()},
\code{\link{HIRMonSetM}()},
\code{\link{HIRMonSetTagTypeRanges}()},
\code{\link{HIRMonSetTagTypes}()},
\code{\link{HIRMonTakeBiasVector}()},
\code{\link{HIRMonTakeData}()},
\code{\link{HIRMonTakeHubItemMatrix}()},
\code{\link{HIRMonTakeItemHubMatrix}()},
\code{\link{HIRMonTakeM01}()},
\code{\link{HIRMonTakeM}()},
\code{\link{HIRMonTakeTagTypeRanges}()},
\code{\link{HIRMonTakeTagTypes}()},
\code{\link{HIRMonTakeValue}()}
}
\concept{Set/Take functions}
