% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonTakeHubItemMatrix}
\alias{HIRMonTakeHubItemMatrix}
\title{Take HubItemMatrix.}
\usage{
HIRMonTakeHubItemMatrix(hirObj, functionName = "HIRMonTakeHubItemMatrix")
}
\arguments{
\item{hirObj}{An HIRMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{HIRMonFailureSymbol}.
}
\description{
Takes HubItemMatrix from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{HIRMonSetBiasVector}()},
\code{\link{HIRMonSetData}()},
\code{\link{HIRMonSetHubItemMatrix}()},
\code{\link{HIRMonSetItemHubMatrix}()},
\code{\link{HIRMonSetM01}()},
\code{\link{HIRMonSetM}()},
\code{\link{HIRMonSetTagTypeRanges}()},
\code{\link{HIRMonSetTagTypes}()},
\code{\link{HIRMonSetValue}()},
\code{\link{HIRMonTakeBiasVector}()},
\code{\link{HIRMonTakeData}()},
\code{\link{HIRMonTakeItemHubMatrix}()},
\code{\link{HIRMonTakeM01}()},
\code{\link{HIRMonTakeM}()},
\code{\link{HIRMonTakeTagTypeRanges}()},
\code{\link{HIRMonTakeTagTypes}()},
\code{\link{HIRMonTakeValue}()}
}
\concept{Set/Take functions}
