% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonTakeValue}
\alias{HIRMonTakeValue}
\title{Take the value in a HIRMon object.}
\usage{
HIRMonTakeValue(hirObj)
}
\arguments{
\item{hirObj}{An HIRMon object.}
}
\value{
Just \code{hirObj$Value}.
}
\description{
Takes the value from HIRMon monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{HIRMonSetBiasVector}()},
\code{\link{HIRMonSetData}()},
\code{\link{HIRMonSetHubItemMatrix}()},
\code{\link{HIRMonSetItemHubMatrix}()},
\code{\link{HIRMonSetM01}()},
\code{\link{HIRMonSetM}()},
\code{\link{HIRMonSetTagTypeRanges}()},
\code{\link{HIRMonSetTagTypes}()},
\code{\link{HIRMonSetValue}()},
\code{\link{HIRMonTakeBiasVector}()},
\code{\link{HIRMonTakeData}()},
\code{\link{HIRMonTakeHubItemMatrix}()},
\code{\link{HIRMonTakeItemHubMatrix}()},
\code{\link{HIRMonTakeM01}()},
\code{\link{HIRMonTakeM}()},
\code{\link{HIRMonTakeTagTypeRanges}()},
\code{\link{HIRMonTakeTagTypes}()}
}
\concept{Set/Take functions}
