% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIRMon.R
\name{HIRMonUnit}
\alias{HIRMonUnit}
\title{Make a HIRMon Unit}
\usage{
HIRMonUnit()
}
\value{
An S3 class "HIRMon". In other words, a list with the attribute "class" set to "HIRMon".
}
\description{
Creates a monad object.
}
