% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubItemDynamicRanks.R
\name{HubItemRanks}
\alias{HubItemRanks}
\title{Hub-item ranks.}
\usage{
HubItemRanks(
  hubsAndItemsMat,
  hubIDs,
  hubTags,
  tagSets,
  tagPresenceFraction = 0.8,
  bias = NULL,
  normalizationFunction = max,
  maxSteps = 100,
  sumFactor = 0.2,
  numberOfHubRanks = 200,
  numberOfItemRanks = 200,
  tol = 10^(-6)
)
}
\arguments{
\item{hubsAndItemsMat}{A bipartite graph adjacency matrix of hub-item connections
with row-names and col-names corresponding to the hub and item ID's.}

\item{hubIDs}{A data frame with columns \code{c('HubID','HubIndex')} mapping ID's to indexes in \code{hubsAndItemsMat}.}

\item{hubTags}{A list of tags lists, each characterizing its corresponding hub in \code{hubIDs}.}

\item{tagSets}{A list of (frequent) metadata sets.}

\item{tagPresenceFraction}{What fraction of each tag set is in \code{hubTags}?}

\item{bias}{A numerical vector with biasing values for each hub and item.}

\item{normalizationFunction}{A normalization function applied to the ranks vector for each tag set.}

\item{tol}{Tolerance for \code{\link{PowerMethod}}.}
}
\value{
A list of sparse arrays corresponding to the item ranks for \code{tagSets}.
}
\description{
Calculate the hub-item ranks for a given adjacency matrix of bi-partite graph and characterizing tags of the nodes.
}
