% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubItemDynamicRanks.R
\name{RemoveZeroEntries}
\alias{RemoveZeroEntries}
\title{Remove zero entries.}
\usage{
RemoveZeroEntries(x, ....)
}
\arguments{
\item{x}{A sparse matrix.}

\item{tol}{A non-negative number to be used as tolerance for checking
if an entry \code{x[i,j]} should be considered to be zero.}
}
\description{
Remove the explicit zero entries from a sparse matrix.
}
\details{
Just calls \code{\link{Matrix::drop0}}.
}
